deployment cleaned and ready to be decoupled if needed
