+++ src/components/Auth/interface.ts
@@ -1,4 +1,5 @@
 import { IUserModel, IUser } from '../User/model';
+import { Request } from 'express';
 
 /**
  * @export
@@ -11,4 +12,6 @@ export interface IAuthService {
      * @memberof AuthService
      */
     findProfileOrCreate(user: IUser): Promise<IUserModel>;
+    deseralizeToken(req: Request): Promise<any>;
+    authUser(req: Request): Promise<IUserModel>;
 }
