+++ src/components/Auth/service.ts
@@ -2,7 +2,10 @@ import { Types } from 'mongoose';
 import { IOrganization } from '../Organization/model';
 import OrganizationService from '../Organization/service';
 import UserModel, { IUserModel, IUser } from '../User/model';
+import UserService from '../User/service';
 import { IAuthService } from './interface';
+import { Request } from 'express';
+import JWTTokenService from '../Session/service';
 
 /**
  * @export
@@ -48,6 +51,32 @@ const AuthService: IAuthService = {
             throw new Error(error);
         }
     },
+
+    async deseralizeToken(req: Request): Promise<any> {
+        const argoDecodedHeaderToken: any = await JWTTokenService.DecodeToken(req);
+        return JWTTokenService.VerifyToken(argoDecodedHeaderToken);
+    },
+
+    async authUser(req: Request): Promise<IUserModel> {
+        const deserializedToken: any = await this.deseralizeToken(req);
+        const user: IUserModel = await UserService.findOne(deserializedToken.session_id);
+    
+        const { orgId }: { orgId: string } = req.body;
+        const organization: IOrganization = await OrganizationService.findOne(orgId);
+    
+        if(!organization) {
+            return null;
+        }
+            
+        const isUserInOrganization: boolean = user.organizations.some((orgUser) => {
+            return orgUser._id.equals(orgId)
+        });
+    
+        if (!isUserInOrganization) {
+            return null;
+        }
+        return user;
+    }
 };
 
 export default AuthService;
