+++ src/components/GitHubApp/interface.ts
@@ -8,4 +8,6 @@ export interface IGitHubAppTokenService {
     create(gitHubToke: IGuHubAppToken): Promise<boolean>;
     findAndCreate(gitHubId: number, token: string, installationId: number): Promise<boolean>;
     remove(installationId: number): Promise<boolean>;
+    createInstallationToken (installationId: any): Promise<any>;
+    getFullGithubUrlAndFolderName(githubUrl: string, isPrivate: boolean, branch: string, installationId: string, repositoryId: string, owner: string): Promise<Array<string>>;
 }
