+++ src/routes/LogsRouter.ts
@@ -3,8 +3,8 @@ import { LogsComponent } from "../components";
 
 const router: Router = Router();
 
-router.post("/", LogsComponent.newDeploy);
+router.post("/", LogsComponent.deploy);
 
-router.get("/:id", LogsComponent.FindDeploymentById);
+router.get("/:id", LogsComponent.findDeploymentById);
 
 export default router;
